agent_eval_lab/requirements.txt
numpy==1.26.4
torch==2.2.2
PyYAML==6.0.1
